formatiranje koda

+++ www/hardshop.hr/app/controller/NarudzbaController.php
@@ -5,50 +5,49 @@ use PhpOffice\PhpSpreadsheet\Writer\Xlsx;
 
 class NarudzbaController extends AutorizacijaController implements ViewSucelje
 {
-    private $viewPutanja='privatno' . DIRECTORY_SEPARATOR . 'narudzbe' . DIRECTORY_SEPARATOR;
+    private $viewPutanja = 'privatno' . DIRECTORY_SEPARATOR . 'narudzbe' . DIRECTORY_SEPARATOR;
     private $nf;
     private $e;
-    private $poruka='';
+    private $poruka = '';
 
     public function __construct()
     {
         parent::__construct();
-        $this->nf=new NumberFormatter('hr-HR',NumberFormatter::DECIMAL);
+        $this->nf = new NumberFormatter('hr-HR', NumberFormatter::DECIMAL);
         $this->nf->setPattern(App::config('formatBroja'));
     }
 
-    public function index()    
+    public function index()
     {
 
-        $this->view->render($this->viewPutanja . 'index',[
-            'podaci'=>$this->prilagodiPodatke(Narudzba::read())
+        $this->view->render($this->viewPutanja . 'index', [
+            'podaci' => $this->prilagodiPodatke(Narudzba::read())
         ]);
-        
     }
 
     private function prilagodiPodatke($narudzbe)
     {
-        foreach($narudzbe as $n){            
-            if($n->datumnarudzbe==null){
-                $n->datumnarudzbe='Nije definirano';
-            }else{
-                $n->datumnarudzbe=date('d.m.Y.',strtotime($n->datumnarudzbe));
+        foreach ($narudzbe as $n) {
+            if ($n->datumnarudzbe == null) {
+                $n->datumnarudzbe = 'Nije definirano';
+            } else {
+                $n->datumnarudzbe = date('d.m.Y.', strtotime($n->datumnarudzbe));
             }
-            if($n->datumisporuke==null){
-                $n->datumisporuke='Nije definirano';
-            }else{
-                $n->datumisporuke=date('d.m.Y.',strtotime($n->datumisporuke));
+            if ($n->datumisporuke == null) {
+                $n->datumisporuke = 'Nije definirano';
+            } else {
+                $n->datumisporuke = date('d.m.Y.', strtotime($n->datumisporuke));
             }
-            if($n->datumplacanja==null){
-                $n->datumplacanja='Nije definirano';
-            }else{
-                $n->datumplacanja=date('d.m.Y.',strtotime($n->datumplacanja));
+            if ($n->datumplacanja == null) {
+                $n->datumplacanja = 'Nije definirano';
+            } else {
+                $n->datumplacanja = date('d.m.Y.', strtotime($n->datumplacanja));
             }
-            if($n->kupac==null){
-                $n->kupac='Nije definirano';
+            if ($n->kupac == null) {
+                $n->kupac = 'Nije definirano';
             }
-            if($n->vrstaplacanja==null){
-                $n->vrstaplacanja='Nije definirano';
+            if ($n->vrstaplacanja == null) {
+                $n->vrstaplacanja = 'Nije definirano';
             }
         }
         return $narudzbe;
@@ -56,40 +55,41 @@ class NarudzbaController extends AutorizacijaController implements ViewSucelje
 
     public function novi()
     {
-        $kupacSifra=Kupac::prviKupac();
-        if($kupacSifra==0){
+        $kupacSifra = Kupac::prviKupac();
+        if ($kupacSifra == 0) {
             header('location: ' . App::config('url') . 'kupac?p=1');
         }
 
-        $placanjeSifra=Placanje::prvoPlacanje();
-        if($placanjeSifra==0){
+        $placanjeSifra = Placanje::prvoPlacanje();
+        if ($placanjeSifra == 0) {
             header('location: ' . App::config('url') . 'placanje?p=1');
         }
 
         $this->promjena(Narudzba::create([
-            'brojnarudzbe'=>'',
-            'datumnarudzbe'=>'',
-            'datumisporuke'=>null,
-            'datumplacanja'=>null,
-            'kupac'=>$kupacSifra,
-            'placanje'=>$placanjeSifra
+            'brojnarudzbe' => '',
+            'datumnarudzbe' => '',
+            'datumisporuke' => null,
+            'datumplacanja' => null,
+            'kupac' => $kupacSifra,
+            'placanje' => $placanjeSifra
         ]));
     }
 
-    public function odustani($sifra='')
+    public function odustani($sifra = '')
     {
-        $e=Narudzba::readOne($sifra);
+        $e = Narudzba::readOne($sifra);
 
-        if(
-        $e->datumisporuke==null &&
-        $e->datumplacanja==null &&
-        $e->proizvod=Narudzba::detaljiNarudzbe($sifra)==null){
-        Narudzba::delete($e->sifra);
+        if (
+            $e->datumisporuke == null &&
+            $e->datumplacanja == null &&
+            $e->proizvod = Narudzba::detaljiNarudzbe($sifra) == null
+        ) {
+            Narudzba::delete($e->sifra);
         }
         header('location: ' . App::config('url') . 'narudzba');
     }
 
-    public function promjena($sifra='')
+    public function promjena($sifra = '')
     {
         parent::setCSSdependency([
             '<link rel="stylesheet" href="' . App::config('url') . 'public/css/dependency/jquery-ui.css">'
@@ -102,28 +102,28 @@ class NarudzbaController extends AutorizacijaController implements ViewSucelje
             </script>'
         ]);
 
-        if($_SERVER['REQUEST_METHOD']==='GET'){
+        if ($_SERVER['REQUEST_METHOD'] === 'GET') {
             $this->promjena_GET($sifra);
             return;
         }
 
-        $this->e=(object)$_POST;
-        $this->e->proizvodi=Narudzba::detaljiNarudzbe($sifra);
+        $this->e = (object)$_POST;
+        $this->e->proizvodi = Narudzba::detaljiNarudzbe($sifra);
 
         try {
-            $odabraniKupac=Kupac::readOne($this->e->kupac);
-            $this->e->sifra=$sifra;
+            $odabraniKupac = Kupac::readOne($this->e->kupac);
+            $this->e->sifra = $sifra;
             $this->kontrola();
             $this->pripremiZaBazu();
             Narudzba::update((array)$this->e);
             header('location:' . App::config('url') . 'narudzba');
         } catch (\Exception $th) {
-            $this->view->render($this->viewPutanja . 'detalji',[
-                'legend'=>'Unos nove narudzbe',
-                'poruka'=>$this->poruka,
-                'kupac'=>$odabraniKupac,
-                'e'=>$this->e,
-                'placanja'=>Placanje::read()
+            $this->view->render($this->viewPutanja . 'detalji', [
+                'legend' => 'Unos nove narudzbe',
+                'poruka' => $this->poruka,
+                'kupac' => $odabraniKupac,
+                'e' => $this->e,
+                'placanja' => Placanje::read()
             ]);
         }
     }
@@ -136,27 +136,27 @@ class NarudzbaController extends AutorizacijaController implements ViewSucelje
 
     private function kontrolaBrojNarudzbe()
     {
-        $s=$this->e->brojnarudzbe;
+        $s = $this->e->brojnarudzbe;
 
-        if(strlen(trim($s))===0){
-            $this->poruka='Broj narudžbe obavezan';
+        if (strlen(trim($s)) === 0) {
+            $this->poruka = 'Broj narudžbe obavezan';
             throw new Exception();
         }
 
-        $brojnarudzbe=$this->nf->parse($this->e->brojnarudzbe);
-        if($brojnarudzbe<=0){
-            $this->poruka='Broj narudžbe mora biti veći od 0';
+        $brojnarudzbe = $this->nf->parse($this->e->brojnarudzbe);
+        if ($brojnarudzbe <= 0) {
+            $this->poruka = 'Broj narudžbe mora biti veći od 0';
             throw new Exception();
         }
 
-        if(isset($this->e->sifra)){
-            if(!Narudzba::postojiIstiBrojNarudzbe($this->e->brojnarudzbe,$this->e->sifra)){
-                $this->poruka='Broj narudžbe već postoji u bazi';
+        if (isset($this->e->sifra)) {
+            if (!Narudzba::postojiIstiBrojNarudzbe($this->e->brojnarudzbe, $this->e->sifra)) {
+                $this->poruka = 'Broj narudžbe već postoji u bazi';
                 throw new Exception();
             }
-        }else{
-            if(!Narudzba::postojiIstiBrojNarudzbe($this->e->brojnarudzbe)){
-                $this->poruka='Broj narudžbe već postoji u bazi';
+        } else {
+            if (!Narudzba::postojiIstiBrojNarudzbe($this->e->brojnarudzbe)) {
+                $this->poruka = 'Broj narudžbe već postoji u bazi';
                 throw new Exception();
             }
         }
@@ -164,49 +164,49 @@ class NarudzbaController extends AutorizacijaController implements ViewSucelje
 
     private function kontrolaDatumNarudzbe()
     {
-        $s=$this->e->datumnarudzbe;
+        $s = $this->e->datumnarudzbe;
 
-        if(strlen(trim($s))===0){
-            $this->poruka='Datum narudžbe obavezan';
+        if (strlen(trim($s)) === 0) {
+            $this->poruka = 'Datum narudžbe obavezan';
             throw new Exception();
         }
     }
 
     private function promjena_GET($sifra)
     {
-        $this->e=Narudzba::readOne($sifra);
-        $placanja=[];
-        $p=new stdClass();
-        $p->sifra=0;
-        $p->vrstaplacanja='Nije odabrano';
-        $placanja[]=$p;
-        foreach(Placanje::read() as $placanje){
-            $placanja[]=$placanje;
+        $this->e = Narudzba::readOne($sifra);
+        $placanja = [];
+        $p = new stdClass();
+        $p->sifra = 0;
+        $p->vrstaplacanja = 'Nije odabrano';
+        $placanja[] = $p;
+        foreach (Placanje::read() as $placanje) {
+            $placanja[] = $placanje;
         }
 
-        if($this->e->datumnarudzbe!=null){
-            $this->e->datumnarudzbe=date('Y-m-d',strtotime($this->e->datumnarudzbe));
+        if ($this->e->datumnarudzbe != null) {
+            $this->e->datumnarudzbe = date('Y-m-d', strtotime($this->e->datumnarudzbe));
         }
-        if($this->e->datumisporuke!=null){
-            $this->e->datumisporuke=date('Y-m-d',strtotime($this->e->datumisporuke));
+        if ($this->e->datumisporuke != null) {
+            $this->e->datumisporuke = date('Y-m-d', strtotime($this->e->datumisporuke));
         }
-        if($this->e->datumplacanja!=null){
-            $this->e->datumplacanja=date('Y-m-d',strtotime($this->e->datumplacanja));
+        if ($this->e->datumplacanja != null) {
+            $this->e->datumplacanja = date('Y-m-d', strtotime($this->e->datumplacanja));
         }
 
-        $this->view->render($this->viewPutanja . 'detalji',[
-            'e'=>$this->e,
-            'kupac'=>Kupac::readOne($this->e->kupac),
-            'placanja'=>$placanja,
-            'legend'=>'',
-            'poruka'=>$this->poruka
+        $this->view->render($this->viewPutanja . 'detalji', [
+            'e' => $this->e,
+            'kupac' => Kupac::readOne($this->e->kupac),
+            'placanja' => $placanja,
+            'legend' => '',
+            'poruka' => $this->poruka
         ]);
     }
 
-    public function brisanje($sifra=0)
+    public function brisanje($sifra = 0)
     {
-        $sifra=(int)$sifra;
-        if($sifra===0){
+        $sifra = (int)$sifra;
+        if ($sifra === 0) {
             header('location: ' . App::config('url') . 'index/odjava');
             return;
         }
@@ -216,133 +216,136 @@ class NarudzbaController extends AutorizacijaController implements ViewSucelje
 
     public function pripremiZaView()
     {
-
     }
 
     public function pripremiZaBazu()
     {
-        if($this->e->kupac==0){
-            $this->e->kupac=null;
+        if ($this->e->kupac == 0) {
+            $this->e->kupac = null;
         }
-        if($this->e->datumnarudzbe==''){
-            $this->e->datumnarudzbe=null;
+        if ($this->e->datumnarudzbe == '') {
+            $this->e->datumnarudzbe = null;
         }
-        if($this->e->datumplacanja==''){
-            $this->e->datumplacanja=null;
+        if ($this->e->datumplacanja == '') {
+            $this->e->datumplacanja = null;
         }
-        if($this->e->datumisporuke==''){
-            $this->e->datumisporuke=null;
+        if ($this->e->datumisporuke == '') {
+            $this->e->datumisporuke = null;
         }
     }
 
     public function pocetniPodaci()
     {
-        $e=new stdClass();
-        $e->brojnarudzbe='';
-        $e->datumnarudzbe='';
-        $e->datumisporuke=null;
-        $e->datumplacanja=null;
-        $e->kupac=0;
-        $e->placanje=null;
+        $e = new stdClass();
+        $e->brojnarudzbe = '';
+        $e->datumnarudzbe = '';
+        $e->datumisporuke = null;
+        $e->datumplacanja = null;
+        $e->kupac = 0;
+        $e->placanje = null;
         return $e;
     }
 
     private function formatIznosa($broj)
     {
-        if($broj==null){
+        if ($broj == null) {
             return $this->nf->format(0);
         }
-            return $this->nf->format($broj);
+        return $this->nf->format($broj);
     }
 
     public function dodajproizvod()
     {
         $res = new stdClass();
-        if(!Narudzba::postojiProizvodNarudzba($_GET['narudzba'],
-                    $_GET['proizvod'])){
-            Narudzba::dodajProizvodNarudzba($_GET['narudzba'],
-                    $_GET['proizvod']);
-            $res->error=false;
-            $res->description='Uspješno dodano';
-                    }else{
-                        $res->error=true;
-                        $res->description='Proizvod već postoji u narudžbi';
-                    }
+        if (!Narudzba::postojiProizvodNarudzba(
+            $_GET['narudzba'],
+            $_GET['proizvod']
+        )) {
+            Narudzba::dodajProizvodNarudzba(
+                $_GET['narudzba'],
+                $_GET['proizvod']
+            );
+            $res->error = false;
+            $res->description = 'Uspješno dodano';
+        } else {
+            $res->error = true;
+            $res->description = 'Proizvod već postoji u narudžbi';
+        }
 
-                    header('Content-Type: application/json; charset=utf-8');
-                    echo json_encode($res,JSON_NUMERIC_CHECK);
+        header('Content-Type: application/json; charset=utf-8');
+        echo json_encode($res, JSON_NUMERIC_CHECK);
     }
 
     public function obrisiproizvod()
     {
-        Narudzba::obrisiProizvodNarudzba($_GET['narudzba'],
-                    $_GET['proizvod']);
-               
+        Narudzba::obrisiProizvodNarudzba(
+            $_GET['narudzba'],
+            $_GET['proizvod']
+        );
     }
 
     public function excel()
     {
 
-    // (B) CREATE A NEW SPREADSHEET
-    $spreadsheet = new Spreadsheet();
-    $sheet = $spreadsheet->getActiveSheet();
-    $sheet->setTitle("Sve narudžbe");
+        // (B) CREATE A NEW SPREADSHEET
+        $spreadsheet = new Spreadsheet();
+        $sheet = $spreadsheet->getActiveSheet();
+        $sheet->setTitle("Sve narudžbe");
 
-    // (C) SET CELL VALUE
+        // (C) SET CELL VALUE
 
-    $sheet->setCellValue('A1', 'Broj narudžbe');
-    $sheet->setCellValue('B1', 'Kupac');
-    $sheet->setCellValue('C1', 'Datum narudžbe');
-    $sheet->setCellValue('D1', 'Vrsta plaćanja');
-    $sheet->setCellValue('E1', 'Naziv proizvoda');
-    $sheet->setCellValue('F1', 'Cijena proizvoda');
+        $sheet->setCellValue('A1', 'Broj narudžbe');
+        $sheet->setCellValue('B1', 'Kupac');
+        $sheet->setCellValue('C1', 'Datum narudžbe');
+        $sheet->setCellValue('D1', 'Vrsta plaćanja');
+        $sheet->setCellValue('E1', 'Naziv proizvoda');
+        $sheet->setCellValue('F1', 'Cijena proizvoda');
 
 
-    $spreadsheet->getActiveSheet()->getStyle('A1:F1')->getFont()->setBold( true );
+        $spreadsheet->getActiveSheet()->getStyle('A1:F1')->getFont()->setBold(true);
 
-    $redniBroj=2;
+        $redniBroj = 2;
 
-     foreach(Narudzba::readExcelNarudzba() as $red){
-         $sheet->setCellValue('A' . $redniBroj, $red->brojnarudzbe);
-         $sheet->setCellValue('B' . $redniBroj, $red->kupac);
-         $sheet->setCellValue('C' . $redniBroj, $red->datumnarudzbe=date('d.m.Y.'));
-         $sheet->setCellValue('D' . $redniBroj, $red->vrstaplacanja);
-         $sheet->setCellValue('E' . $redniBroj, $red->naziv);
-         $sheet->setCellValue('F' . $redniBroj, $red->cijena);
+        foreach (Narudzba::readExcelNarudzba() as $red) {
+            $sheet->setCellValue('A' . $redniBroj, $red->brojnarudzbe);
+            $sheet->setCellValue('B' . $redniBroj, $red->kupac);
+            $sheet->setCellValue('C' . $redniBroj, $red->datumnarudzbe);
+            $sheet->setCellValue('D' . $redniBroj, $red->vrstaplacanja);
+            $sheet->setCellValue('E' . $redniBroj, $red->naziv);
+            $sheet->setCellValue('F' . $redniBroj, $red->cijena);
 
-         $redniBroj++;
-     }
-     $sheet->setCellValue('F' . $redniBroj, '=sum(F2:F' . ($redniBroj-1) . ')');
+            $redniBroj++;
+        }
+        $sheet->setCellValue('F' . $redniBroj, '=sum(F2:F' . ($redniBroj - 1) . ')');
 
-     foreach ($sheet->getColumnIterator() as $column) {
-        $sheet->getColumnDimension($column->getColumnIndex())->setAutoSize(true);
-     }
+        foreach ($sheet->getColumnIterator() as $column) {
+            $sheet->getColumnDimension($column->getColumnIndex())->setAutoSize(true);
+        }
 
-     $spreadsheet->getActiveSheet()
-     ->getStyle('F2:F'. $redniBroj)
-     ->getNumberFormat()
-     ->setFormatCode(PhpOffice\PhpSpreadsheet\Style\NumberFormat::FORMAT_CURRENCY_EUR);
-     $spreadsheet->getActiveSheet()
-    ->getStyle('A1:F' . $redniBroj)
-    ->getBorders()
-    ->getOutline()
-    ->setBorderStyle(PhpOffice\PhpSpreadsheet\Style\Border::BORDER_THICK);
-    
+        $spreadsheet->getActiveSheet()
+            ->getStyle('F2:F' . $redniBroj)
+            ->getNumberFormat()
+            ->setFormatCode(PhpOffice\PhpSpreadsheet\Style\NumberFormat::FORMAT_CURRENCY_EUR);
+        $spreadsheet->getActiveSheet()
+            ->getStyle('A1:F' . $redniBroj)
+            ->getBorders()
+            ->getOutline()
+            ->setBorderStyle(PhpOffice\PhpSpreadsheet\Style\Border::BORDER_THICK);
 
-    // (D) SEND DOWNLOAD HEADERS
-    // ob_clean();
-    // ob_start();
-    $writer = new Xlsx($spreadsheet);
-    header('Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
-    header('Content-Disposition: attachment;filename="Narudžbe.xlsx"');
-    header('Cache-Control: max-age=0');
-    header('Expires: Fri, 11 Nov 2011 11:11:11 GMT');
-    header('Last-Modified: '. gmdate('D, d M Y H:i:s') .' GMT');
-    header('Cache-Control: cache, must-revalidate');
-    header('Pragma: public');
-    $writer->save('php://output');
-    // ob_end_flush();
+
+        // (D) SEND DOWNLOAD HEADERS
+        // ob_clean();
+        // ob_start();
+        $writer = new Xlsx($spreadsheet);
+        header('Content-Type: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet');
+        header('Content-Disposition: attachment;filename="Narudžbe.xlsx"');
+        header('Cache-Control: max-age=0');
+        header('Expires: Fri, 11 Nov 2011 11:11:11 GMT');
+        header('Last-Modified: ' . gmdate('D, d M Y H:i:s') . ' GMT');
+        header('Cache-Control: cache, must-revalidate');
+        header('Pragma: public');
+        $writer->save('php://output');
+        // ob_end_flush();
 
     }
-
 }
